rm op_function_common
